+++ src/lhp.ts
@@ -1,4 +1,4 @@
-#!/usr/bin/env bun
+#!/usr/bin/env node
 import { main } from './cli';
 
 main(process.argv.slice(2)).catch((err: Error) => {